closeReason utáni változások backenden

+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/CloseReasonRepository.java
@@ -3,5 +3,5 @@ package com.Hbence.appointmentManagementAPI.repository;
 import com.Hbence.appointmentManagementAPI.entity.CloseReason;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CloseReasonRepository extends JpaRepository<CloseReason, Long> {
+public interface CloseReasonRepository extends JpaRepository<CloseReason, Integer> {
 }
